Enable WebmFileWriter/WebmFileMuxer classes.

+++ source/base/codec/webm_file_muxer.cc
@@ -169,7 +169,7 @@ bool WebmFileMuxer::writeFrame(std::string_view frame,
                                uint64_t track_num, bool is_key)
 {
     if (!segment_->AddFrame(reinterpret_cast<const uint8_t*>(frame.data()), frame.size(),
-                            track_num, timestamp.count(), is_key))
+                            track_num, static_cast<uint64_t>(timestamp.count()), is_key))
     {
         LOG(LS_ERROR) << "AddFrame failed";
         return false;
